# packages/types/openapi.yaml
openapi: 3.1.0
info:
  title: NEXUS.SALES API
  version: 0.1.0
servers:
  - url: http://localhost:3001
paths:
  /api/auth/register:
    post:
      summary: Register seller
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSuccess" }
        "409":
          description: Email exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/auth/login:
    post:
      summary: Login seller
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSuccess" }
        "400":
          description: Invalid creds
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      properties:
        success: { type: boolean, const: false }
        message: { type: string }
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthSuccess:
      type: object
      properties:
        success: { type: boolean, const: true }
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        createdAt: { type: string, format: date-time }
...